+++ R/simulate_MDP.R
@@ -20,9 +20,10 @@
 #' @param n number of trajectories.
 #' @param start probability distribution over the states for choosing the
 #'  starting states for the trajectories. Defaults to "uniform".
-#' @param horizon number of epochs for the simulation. If `NULL` then the
+#' @param horizon epochs end once an absorbing state is reached or after 
+#'  the maximal number of epochs specified via `horizon`. If `NULL` then the
 #'  horizon for the model is used.
-#' @param return_states logical; return visited states.
+#' @param return_trajectories logical; return the complete trajectories.
 #' @param epsilon the probability of random actions  for using an epsilon-greedy policy.
 #'  Default for solved models is 0 and for unsolved model 1.
 #' @param engine `'cpp'` or `'r'` to perform simulation using a faster C++
@@ -35,32 +36,35 @@
 #'  * `reward`: Reward for each trajectory.
 #'  * `action_cnt`: Action counts.
 #'  * `state_cnt`: State counts.
-#'  * `states`: a vector with state ids.
-#'    Rows represent trajectories.
+#'  * `trajectories`: A data.frame with the trajectories. Each row 
+#'    contains the `episode` id, the `time` step, the state `s`, 
+#'    the chosen action `a`,
+#'    the reward `r`, and the next state `s_prime`. Trajectories are 
+#'    only returend of `return_trajectories = TRUE`.
 #' @author Michael Hahsler
-#' @return A vector with state ids (in the final epoch or all). Attributes containing action
-#' counts, and rewards  for each trajectory may be available.
-#' @author Michael Hahsler
-#' @md
 #' @examples
+#' # enable parallel simulation 
+#' # doParallel::registerDoParallel()
+#' 
 #' data(Maze)
 #'
 #' # solve the POMDP for 5 epochs and no discounting
 #' sol <- solve_MDP(Maze, discount = 1)
 #' sol
-#' 
+#'
 #' # U in the policy is and estimate of the utility of being in a state when using the optimal policy.
 #' policy(sol)
 #' matrix(policy(sol)$action, nrow = 3, dimnames = list(1:3, 1:4))[3:1, ]
 #'
-#' ## Example 1: simulate 10 trajectories following the policy, only the final belief state is returned
+#' ## Example 1: simulate 100 trajectories following the policy, 
+#' #             only the final belief state is returned
 #' sim <- simulate_MDP(sol, n = 100, horizon = 10, verbose = TRUE)
 #' sim
-#' 
-#' # Note that all simulations start at s_1 and that the simulated avg. reward 
+#'
+#' # Note that all simulations start at s_1 and that the simulated avg. reward
 #' # is therefore an estimate to the U value for the start state s_1.
-#' policy(sol)[1,] 
-#' 
+#' policy(sol)[1,]
+#'
 #' # Calculate proportion of actions taken in the simulation
 #' round_stochastic(sim$action_cnt / sum(sim$action_cnt), 2)
 #'
@@ -68,25 +72,28 @@
 #' hist(sim$reward)
 #'
 #' ## Example 2: simulate starting following a uniform distribution over all
-#' #             states and return all visited states
-#' sim <- simulate_MDP(sol, n = 100, start = "uniform", horizon = 10, return_states = TRUE)
-#' sim$avg_reward
+#' #             states and return all trajectories
+#' sim <- simulate_MDP(sol, n = 100, start = "uniform", horizon = 10, 
+#'   return_trajectories = TRUE)
+#' head(sim$trajectories)   
+#'   
 #'
 #' # how often was each state visited?
-#' table(sim$states)
-#' matrix(table(sim$states), nrow = 3, dimnames = list(1:3, 1:4))[3:1, ]
+#' table(sim$trajectories$s)
+#' matrix(table(sim$trajectories$s), nrow = 3, 
+#'   dimnames = list(1:3, 1:4))[3:1, ]
 #' @export
 simulate_MDP <-
   function(model,
-    n = 100,
-    start = NULL,
-    horizon = NULL,
-    return_states = FALSE,
-    epsilon = NULL,
-    delta_horizon = 1e-3,
-    engine = "cpp",
-    verbose = FALSE,
-    ...) {
+           n = 100,
+           start = NULL,
+           horizon = NULL,
+           return_trajectories = FALSE,
+           epsilon = NULL,
+           delta_horizon = 1e-3,
+           engine = "cpp",
+           verbose = FALSE,
+           ...) {
     engine <- match.arg(tolower(engine), c("cpp", "r"))
     
     start <- .translate_belief(start, model = model)
@@ -97,13 +104,15 @@ simulate_MDP <-
     if (is.null(horizon))
       horizon <- model$horizon
     if (is.null(horizon) || is.infinite(horizon)) {
-      if (is.null(model$discount) || !(model$discount < 1)) 
+      if (is.null(model$discount) || !(model$discount < 1))
         stop("Simulation needs a finite simulation horizon.")
       
-      # find a horizon that approximates the reward using 
+      # find a horizon that approximates the reward using
       # discount^horizon * max_abs_R <= 0.001
-      max_abs_R <-  max(abs(reward_matrix(model, sparse = TRUE)$value))
-      horizon <- ceiling(log(delta_horizon/max_abs_R)/log(model$discount))
+      max_abs_R <-
+        max(abs(reward_matrix(model, sparse = TRUE)$value))
+      horizon <-
+        ceiling(log(delta_horizon / max_abs_R) / log(model$discount))
     }
     horizon <- as.integer(horizon)
     
@@ -123,7 +132,7 @@ simulate_MDP <-
     
     if (engine == "cpp") {
       ### TODO: Make sparse
-      model <- normalize_MDP(model, sparse = FALSE)
+      model <- normalize_MDP(model, sparse = TRUE)
       
       if (foreach::getDoParWorkers() == 1 || n * horizon < 100000)
         return (simulate_MDP_cpp(
@@ -132,9 +141,9 @@ simulate_MDP <-
           start,
           horizon,
           disc,
-          return_states,
+          return_trajectories,
           epsilon,
-          verbose
+          verbose = verbose
         ))
       
       ns <- foreach_split(n)
@@ -154,23 +163,30 @@ simulate_MDP <-
       
       sim <- foreach(w = 1:length(ns)) %dopar%
         simulate_MDP_cpp(model,
-          ns[w],
-          start,
-          horizon,
-          disc,
-          return_states,
-          epsilon,
-          verbose = FALSE)
+                         ns[w],
+                         start,
+                         horizon,
+                         disc,
+                         return_trajectories,
+                         epsilon,
+                         verbose = FALSE)
+      
+      # adjust the episode number for parallel processing
+      episode_add <- cumsum(c(0L, ns))
+      for (i in seq_along(sim)) {
+        sim[[i]]$trajectories$episode <-
+          sim[[i]]$trajectories$episode + episode_add[i]
+      }
       
       rew <- Reduce(c,  lapply(sim, "[[", "reward"))
-      
+     
       return(
         list(
           avg_reward = mean(rew, na.rm = TRUE),
           reward = rew,
           action_cnt = Reduce('+', lapply(sim, "[[" , "action_cnt")),
           state_cnt =  Reduce('+', lapply(sim, "[[", "state_cnt")),
-          states = Reduce(c, lapply(sim, "[[", "states"))
+          trajectories = Reduce(rbind, lapply(sim, "[[", "trajectories"))
         )
       )
       
@@ -179,6 +195,7 @@ simulate_MDP <-
     ## R implementation starts here
     states <- as.character(model$states)
     n_states <- length(states)
+    states_absorbing <- which(absorbing_states(model))
     actions <- as.character(model$actions)
     
     trans_m <- transition_matrix(model, sparse = NULL)
@@ -197,10 +214,10 @@ simulate_MDP <-
       cat("- engine:", engine, "\n")
       cat("- horizon:", horizon, "\n")
       cat("- n:",
-        n,
-        "- parallel workers:",
-        foreach::getDoParWorkers(),
-        "\n")
+          n,
+          "- parallel workers:",
+          foreach::getDoParWorkers(),
+          "\n")
       cat("- epsilon:", epsilon, "\n")
       cat("- discount factor:", disc, "\n")
       cat("\n")
@@ -210,7 +227,7 @@ simulate_MDP <-
     #sim <- replicate(n, expr = {
     sim <- foreach(i = 1:n) %dopar% {
       # find a initial state
-      s <- sample(states, 1, prob = start)
+      s <- sample.int(length(states), 1L, prob = start)
       
       action_cnt <- rep(0L, length(actions))
       names(action_cnt) <- actions
@@ -218,10 +235,17 @@ simulate_MDP <-
       names(state_cnt) <- states
       rew <- 0
       
-      if (return_states)
-        states_visited <- integer(horizon)
+      if (return_trajectories)
+        trajectory <- data.frame(
+          episode = rep(NA_integer_, horizon),
+          time = rep(NA_integer_, horizon),
+          s = NA_integer_,
+          a = NA_integer_,
+          r = NA_real_,
+          s_prime = NA_integer_
+        )
       else
-        states_visited <- integer()
+        trajectory <- NULL
       
       for (j in seq_len(horizon)) {
         if (runif(1) < epsilon) {
@@ -235,39 +259,67 @@ simulate_MDP <-
         
         s_prev <- s
         s <-
-          sample.int(length(states), 1L, prob = trans_m[[a]][s, ])
+          sample.int(length(states), 1L, prob = trans_m[[a]][s,])
         
         #rew <- rew + rew_m[[a]][[s_prev]][s] * disc ^ (j - 1L)
         # MDPs have no observation!
-        rew <-
-          rew + reward_val(model, a, s_prev, s) * disc ^ (j - 1L)
+        r <- reward_val(model, a, s_prev, s)
+        rew <- rew + r * disc ^ (j - 1L)
         
-        if (return_states)
-          states_visited[j] <- s
-      }
+        if (return_trajectories)
+          trajectory[j, ] <-
+          data.frame(
+            episode = i,
+            time = j - 1L,
+            s = s_prev,
+            a = a,
+            r = r,
+            s_prime = s
+          )
       
-      states_visited <-
-        factor(states_visited,
-          levels = seq_along(model$states),
-          labels = model$states)
+        if(s %in% states_absorbing) {
+          if (return_trajectories)
+            trajectory <- trajectory[1:j, , drop = FALSE]
+            # TODO: maybe add the finals state
+          break
+        }
+        
+      }
       
       list(
         action_cnt =  action_cnt,
         state_cnt = state_cnt,
         reward = rew,
-        states = states_visited
+        trajectory = trajectory
       )
     }
-    #  , simplify = FALSE)
     
     rew <- Reduce(c, lapply(sim, "[[", "reward"))
     rew <- unname(rew)
+    
+    trajectories <- NULL
+    if (return_trajectories) {
+      trajectories <- Reduce(rbind, lapply(sim, "[[", "trajectory"))
+      trajectories$s <-
+        factor(trajectories$s,
+               levels = seq_along(states),
+               labels = states)
+      trajectories$a <-
+        factor(trajectories$a,
+               levels = seq_along(actions),
+               labels = actions)
+      trajectories$s_prime <-
+        factor(trajectories$s_prime,
+               levels = seq_along(states),
+               labels = states)
+    }
+    
     
     list(
       avg_reward = mean(rew, na.rm = TRUE),
       reward = rew,
       action_cnt = Reduce('+', lapply(sim, "[[", "action_cnt")),
       state_cnt = Reduce('+', lapply(sim, "[[", "state_cnt")),
-      states = Reduce(c, lapply(sim, "[[", "states"))
+      trajectories = trajectories
     )
   }